requirements.txt
asgiref==3.8.1
click==8.1.7
cmake==3.30.2
Django==5.1
djangorestframework==3.15.2
dlib==19.24.5
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
sqlparse==0.5.1
